+++ src/gpu/ops/RRectDrawOp.cpp
@@ -76,7 +76,6 @@ void RRectDrawOp::onDraw(RenderPass* renderPass) {
   renderPass->setVertexBuffer(0, vertexBuffer->gpuBuffer(), vertexBufferProxyView->offset());
   renderPass->setIndexBuffer(indexBuffer->gpuBuffer());
   auto numIndicesPerRRect = hasStroke ? IndicesPerStrokeRRect : IndicesPerFillRRect;
-  renderPass->drawIndexed(PrimitiveType::Triangles,
-                          static_cast<uint32_t>(rectCount * numIndicesPerRRect));
+  renderPass->drawIndexed(PrimitiveType::Triangles, rectCount * numIndicesPerRRect);
 }
 }  // namespace tgfx
